refactor(package): 패키지 구조 변경 - repository, entity, exception 폴더로 분류
