+++ src/main/java/qna/domain/entity/ContentType.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.entity;
 
 public enum ContentType {
     QUESTION, ANSWER;
