+++ src/main/java/qna/exception/CannotDeleteException.java
@@ -1,4 +1,4 @@
-package qna;
+package qna.exception;
 
 public class CannotDeleteException extends Exception {
     private static final long serialVersionUID = 1L;
