+++ src/main/java/qna/repository/QuestionRepository.java
@@ -1,6 +1,7 @@
-package qna.domain;
+package qna.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.entity.Question;
 
 import java.util.List;
 import java.util.Optional;
