+++ src/main/java/qna/service/QnAService.java
@@ -4,9 +4,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import qna.CannotDeleteException;
-import qna.NotFoundException;
-import qna.domain.*;
+import qna.exception.CannotDeleteException;
+import qna.exception.NotFoundException;
+import qna.domain.entity.*;
+import qna.repository.AnswerRepository;
+import qna.repository.QuestionRepository;
 
 import javax.annotation.Resource;
 import java.time.LocalDateTime;
