+++ src/test/java/qna/domain/UserTest.java
@@ -1,5 +1,7 @@
 package qna.domain;
 
+import qna.domain.entity.User;
+
 public class UserTest {
     public static final User JAVAJIGI = new User(1L, "javajigi", "password", "name", "javajigi@slipp.net");
     public static final User SANJIGI = new User(2L, "sanjigi", "password", "name", "sanjigi@slipp.net");
